+++ functions/src/index.ts
@@ -13,3 +13,4 @@ export { onNewMessage } from "./onNewMessage";
 export { unmatchUser } from "./unmatchUser";
 export { createStripeCheckout } from "./createStripeCheckout";
 export { stripeWebhook } from "./stripeWebhook";
+export { recordSwipe } from "./recordSwipe";